requirements.txt
pandas
Pillow
ipykernel